CodeChefDay5question54
Chef appeared for an exam consisting of 33 sections. Each section is worth 100100 marks.

Chef scored AA marks in Section 11, BB marks in section 22, and CC marks in section 33.

Chef passes the exam if both of the following conditions satisfy:

Total score of Chef is \geq 100≥100;
Score of each section \geq 10≥10.
Determine whether Chef passes the exam or not.