CodeChefDay5question55
Chef is shopping for masks. In the shop, he encounters 22 types of masks:

Disposable Masks — cost XX but last only 11 day.
Cloth Masks — cost YY but last 1010 days.
Chef wants to buy masks to last him 100100 days. He will buy the masks which cost him the least. In case there is a tie in terms of cost, Chef will be eco-friendly and choose the cloth masks. Which type of mask will Chef choose?