CodeChefDay7question75
Alice has a bucket of water initially having WW litres of water in it. The maximum capacity of the bucket is XX liters.

Alice turned on the tap and the water starts flowing into the bucket at a rate of YY litres/hour. She left the tap running for exactly ZZ hours. Determine whether the bucket has been overflown, filled exactly, or is still left unfilled.